Extract class into own file to prevent parent class from having too many lines
